+++ R/tree_length.R
@@ -23,14 +23,16 @@
 #' tree <- TreeTools::BalancedTree(inapplicable.phyData[[1]])
 #' TreeLength(tree, inapplicable.phyData[[1]])
 #' TreeLength(tree, inapplicable.phyData[[1]], concavity = 10)
+#' \donttest{ # PrepareDataProfile() and random-tree scoring are slower:
 #' TreeLength(tree, inapplicable.phyData[[1]], concavity = "profile")
 #' TreeLength(5, inapplicable.phyData[[1]])
-#' 
+#'
 #' # HSJ scoring with a character hierarchy
 #' dataset6 <- inapplicable.phyData[["Vinther2008"]]
 #' hier <- CharacterHierarchy("1" = 2:3)
 #' tree6 <- TreeTools::BalancedTree(dataset6)
 #' TreeLength(tree6, dataset6, hierarchy = hier, inapplicable = "hsj")
+#' }
 #' @seealso 
 #' - Conduct tree search using [`MaximizeParsimony()`] (command line) or
 #' [`EasyTrees()`] (graphical user interface).
@@ -199,7 +201,7 @@ TreeLength.phylo <- function(tree, dataset, concavity = Inf,
     tip_data <- matrix(unlist(dataset, use.names = FALSE),
                        nrow = length(dataset), byrow = TRUE)
     ts_fitch_score(tree[["edge"]], contrast, tip_data,
-                   at$weight, at$levels)
+                   .ScaleWeight(at$weight), at$levels)
   }
 }
 
@@ -308,7 +310,7 @@ TreeLength.list <- function(tree, dataset, concavity = Inf,
   contrast <- at$contrast
   tip_data <- matrix(unlist(dataset, use.names = FALSE),
                      nrow = length(dataset), byrow = TRUE)
-  weight <- at$weight
+  weight <- .ScaleWeight(at$weight)
   levels <- at$levels
 
   min_steps <- if (iw) as.integer(at[["min.length"]]) else integer(0)
@@ -498,7 +500,8 @@ FastCharacterLength <- function(tree, dataset) {
   }
   tip_data <- matrix(unlist(dataset, use.names = FALSE),
                      nrow = length(dataset), byrow = TRUE)
-  ts_char_steps(tree[["edge"]], at$contrast, tip_data, at$weight, at$levels)
+  ts_char_steps(tree[["edge"]], at$contrast, tip_data,
+                .ScaleWeight(at$weight), at$levels)
 }
 
 #' Calculate parsimony score from Morphy object